fix credo warnings
